+++ server/src/main/java/org/opensearch/action/admin/cluster/stats/TransportClusterStatsAction.java
@@ -172,7 +172,8 @@ public class TransportClusterStatsAction extends TransportNodesAction<
             false,
             false,
             false,
-            false
+            false,
+            true
         );
         List<ShardStats> shardsStats = new ArrayList<>();
         for (IndexService indexService : indicesService) {