+++ test/framework/src/main/java/org/opensearch/cluster/MockInternalClusterInfoService.java
@@ -124,7 +124,8 @@ public class MockInternalClusterInfoService extends InternalClusterInfoService {
                 nodeStats.getSearchPipelineStats(),
                 nodeStats.getSegmentReplicationRejectionStats(),
                 nodeStats.getRepositoriesStats(),
-                nodeStats.getAdmissionControlStats()
+                nodeStats.getAdmissionControlStats(),
+                nodeStats.getPeerRecoveryStats()
             );
         }).collect(Collectors.toList());
     }